+++ tencentcloud/apigateway/v20180808/models.go
@@ -3727,6 +3727,12 @@ type DomainSetList struct {
 
 	// 网络类型（'INNER' 或 'OUTER'）。
 	NetType *string `json:"NetType,omitempty" name:"NetType"`
+
+	// 是否将HTTP请求强制跳转 HTTPS，默认为false。参数为 true时，API网关会将所有使用该自定义域名的 HTTP 协议的请求重定向至 HTTPS 协议进行转发。
+	IsForcedHttps *bool `json:"IsForcedHttps,omitempty" name:"IsForcedHttps"`
+
+	// 域名备案注册状态
+	RegistrationStatus *bool `json:"RegistrationStatus,omitempty" name:"RegistrationStatus"`
 }
 
 type DomainSets struct {