+++ tencentcloud/cme/v20191029/models.go
@@ -149,7 +149,7 @@ type AudioTrackItem struct {
 	// <ul>
 	// <li>VOD ：素材来源于云点播文件 ；</li>
 	// <li>CME ：视频来源于制作云媒体文件 ；</li>
-	// <li>EXTERNAL ：视频来源于媒资绑定。</li>
+	// <li>EXTERNAL ：视频来源于媒资绑定，如果媒体不是存储在腾讯云点播中或者云创中，都需要使用媒资绑定。</li>
 	// </ul>
 	SourceType *string `json:"SourceType,omitempty" name:"SourceType"`
 
@@ -159,6 +159,8 @@ type AudioTrackItem struct {
 	// <li>当 SourceType 为 CME 时，为制作云的媒体 ID，项目归属者必须对该云媒资有访问权限；</li>
 	// <li>当 SourceType 为 EXTERNAL 时，为媒资绑定的 Definition 与 MediaKey 中间用冒号分隔合并后的字符串，格式为 Definition:MediaKey 。</li>
 	// </ul>
+	// 
+	// 注：当 SourceType 为 EXTERNAL 时，目前仅支持外部 URL 的媒体直接导入项目中。当外部 URL Scheme 为 https 时，Definiton 为 1000000，MediaKey 为 URL 去掉 'https://'；当外部 URL Scheme 为 http 时，Definiton 为 1000001，MediaKey 为 URL 去掉 'http://'。
 	SourceMedia *string `json:"SourceMedia,omitempty" name:"SourceMedia"`
 
 	// 音频片段取自媒体文件的起始时间，单位为秒。0 表示从媒体开始位置截取。默认为0。
@@ -2127,10 +2129,14 @@ func (r *ExportVideoEditProjectResponse) FromJsonString(s string) error {
 
 type ExternalMediaInfo struct {
 
-	// 媒资绑定模板 ID。
+	// 媒资绑定模板 ID，可取值为：
+	// <li>1000000：媒体文件为 URL，且 URL Scheme 为 https；</li>
+	// <li>1000001：媒体文件为 URL，且 URL Scheme 为 http。</li>
+	// 
+	// 注：如果要支持其它存储平台或者类型的媒体绑定，请联系 [客服](https://cloud.tencent.com/online-service?from=doc_1156)。
 	Definition *int64 `json:"Definition,omitempty" name:"Definition"`
 
-	// 媒资绑定媒体路径或文件 ID。
+	// 媒资绑定媒体路径或文件 ID，如果要绑定 URL 类型的媒体，请将 URL 的 'https://' 或者 'http://' 去掉。
 	MediaKey *string `json:"MediaKey,omitempty" name:"MediaKey"`
 }
 
@@ -2444,7 +2450,8 @@ type ImportMaterialRequest struct {
 	// 导入媒资类型，取值：
 	// <li>VOD：云点播文件；</li>
 	// <li>EXTERNAL：媒资绑定。</li>
-	// 注意：如果不填默认为云点播文件。
+	// 
+	// 注意：如果不填默认为云点播文件，如果媒体存储在非腾讯云点播中，都需要使用媒资绑定。
 	SourceType *string `json:"SourceType,omitempty" name:"SourceType"`
 
 	// 云点播媒资 FileId，仅当 SourceType 为 VOD 时有效。
@@ -2529,7 +2536,8 @@ type ImportMediaToProjectRequest struct {
 	// 导入媒资类型，取值：
 	// <li>VOD：云点播文件；</li>
 	// <li>EXTERNAL：媒资绑定。</li>
-	// 注意：如果不填默认为云点播文件。
+	// 
+	// 注意：如果不填默认为云点播文件，如果媒体存储在非腾讯云点播中，都需要使用媒资绑定。
 	SourceType *string `json:"SourceType,omitempty" name:"SourceType"`
 
 	// 云点播媒资文件 Id，当 SourceType 取值 VOD 或者缺省的时候必填。
@@ -4155,7 +4163,7 @@ type VideoTrackItem struct {
 	// <ul>
 	// <li>VOD ：媒体来源于云点播文件 。</li>
 	// <li>CME ：视频来源制作云媒体文件。</li>
-	// <li>EXTERNAL ：视频来源于媒资绑定。</li>
+	// <li>EXTERNAL ：视频来源于媒资绑定，如果媒体不是存储在腾讯云点播中或者云创中，都需要使用媒资绑定。</li>
 	// </ul>
 	SourceType *string `json:"SourceType,omitempty" name:"SourceType"`
 
@@ -4165,6 +4173,8 @@ type VideoTrackItem struct {
 	// <li>当 SourceType 为 CME 时，为制作云的媒体 ID，项目归属者必须对该云媒资有访问权限；</li>
 	// <li>当 SourceType 为 EXTERNAL 时，为媒资绑定的 Definition 与 MediaKey 中间用冒号分隔合并后的字符串，格式为 Definition:MediaKey 。</li>
 	// </ul>
+	// 
+	// 注：当 SourceType 为 EXTERNAL 时，目前仅支持外部 URL 的媒体直接导入项目中。当外部 URL Scheme 为 https 时，Definiton 为 1000000，MediaKey 为 URL 去掉 'https://'；当外部 URL Scheme 为 http 时，Definiton 为 1000001，MediaKey 为 URL 去掉 'http://'。
 	SourceMedia *string `json:"SourceMedia,omitempty" name:"SourceMedia"`
 
 	// 视频片段取自媒体文件的起始时间，单位为秒。默认为0。